+++ src/main/java/Application/Repos/HighestGrossingMoviesRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface HighestGrossingMoviesRepo extends JpaRepository<HighestGrossingMovies, Long> {
+public interface HighestGrossingMoviesRepo extends JpaRepository<HighestGrossingMovies, Integer> {
 }
